Restructure core package
