+++ .changeset/892c0be4/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/core", "type": "patch" }], "dependents": [] }
