+++ .changeset/892c0be4/changes.md
@@ -1,0 +1,1 @@
+- Restructure internal code
